adds all endpoint implementations

+++ internal/pkg/beverage/beverage.go
@@ -4,5 +4,5 @@ package beverage
 type Beer struct {
 	Name        string
 	Description string
-	Price       float64
+	Price       uint
 }
